Simplify log writing method; move Enum into DTO namespace
