+++ src/Parser/TestHook/Passed.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Paraunit\Parser\TestHook;
 
-use Paraunit\Parser\TestStatus;
+use Paraunit\Parser\DTO\TestStatus;
 use PHPUnit\Event\Test\Passed as PassedEvent;
 use PHPUnit\Event\Test\PassedSubscriber;
