+++ src/Parser/TestHook/Warning.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Paraunit\Parser\TestHook;
 
-use Paraunit\Parser\TestStatus;
+use Paraunit\Parser\DTO\TestStatus;
 use PHPUnit\Event\Test\PassedWithWarning;
 use PHPUnit\Event\Test\PassedWithWarningSubscriber;
